% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra Algorithm.}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A data frame consists of three variables.}

\item{init_node}{A numeric scalar variable.}
}
\value{
Returns the shortest path to every other node from the starting node as a vector.
}
\description{
The function is an implementation of algorithm for finding the shortest paths between nodes in a graph,
which may represent, for example, road networks.
}
\examples{
dijkstra(wiki_graph, 1)

}
\references{
\href{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}{Dijkstra Algorithm}
}
