% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean algorithm for GCD.}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{numeric scalars of integers.}

\item{y}{numeric scalars of integers.}
}
\value{
If input data is numeric or integer then 
the output will be a integer otheriwise it will return an error.
}
\description{
The function returns the greatest common divisor (GCD) of two numbers.
}
\examples{
euclidean(100, 1000)

}
\references{
\href{https://en.wikipedia.org/wiki/Euclideanalgorithm}{Euclideanalgorithm}
}
