\name{RCourseLab-package}
\alias{RCourseLab-package}
\alias{RCourseLab}
\docType{package}
\title{
\packageTitle{RCourseLab}
}
\description{
\packageDescription{RCourseLab}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RCourseLab}
\packageIndices{RCourseLab}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{RCourseLab}

Maintainer: \packageMaintainer{RCourseLab}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
