% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCourseLab.R
\docType{package}
\name{RCourseLab}
\alias{RCourseLab}
\alias{RCourseLab-package}
\title{RCourseLab: A package for Lab 3 Assigments.}
\description{
The RCourseLab package provides two functions:
euclidean & dijkstra.
}
\section{euclidean functions}{

Euclidean() returns the greatest common divisor (GCD) of two numbers.
}

\section{dijkstra functions}{

dijkstra() returns the shortest path based on the algorithm.
}

