% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wiki_Graph}
\format{An object}
\source{
From the lab manual
}
\usage{
data(wiki_graph)
}
\description{
Wiki_Graph is a graph which consists of several vertices. Here its a made of a dataframe.
This dataframe consists of three variables (v1, v2 and w) that contains the edges of the graph (fromv1 to v2) with the weight of the edge (w).
}
\examples{
dijkstra(graph = wiki_graph, init_node = 1)
}
\references{
\href{http://www.r-tutor.com/r-introduction/data-frame}{R tutorial}
}
\keyword{datasets}
