\name{RCourseLab-package}
\alias{RCourseLab-package}
\docType{package}
\title{
\packageTitle{RCourseLab}
}
\description{
\packageDescription{RCourseLab}
}
\details{

The DESCRIPTION file
\packageDESCRIPTION{RCourseLab}
\packageIndices{RCourseLab}
If input data is numeric or integer then 
the output will be a integer otheriwise it will return an error.
Euclidean() returns the greatest common divisor (GCD) of two numbers.
}
\author{ packageAuthor {Farhana Chowdhury Tondra}

Maintainer packageMaintainer{Farhana Chowdhury Tondra}
}
\references{
https://en.wikipedia.org/wiki/Euclidean algorithm.
}

\keyword{ GCD, Euclidean Algorithm }
\seealso{

\code{\link[<pkg>:<pkg>-package]{<pkg>}}}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
