\name{RCourseLab-package}
\docType{package}
\title{
\packageTitle{RCourseLab}
}
\description{
\packageDescription{RCourseLab}
}
\details{

The DESCRIPTION file
\packageDESCRIPTION{RCourseLab}
\packageIndices{RCourseLab}
Package contains of implementations of two algorithm. Dijkstra is for finding the shortest path from the initial node and Euclidean is for calculating greatest common divisior of two numbers. 
}
\author{ packageAuthor {Farhana Chowdhury Tondra}

Maintainer packageMaintainer{Farhana Chowdhury Tondra}
}
\references{
https://en.wikipedia.org/wiki/Euclideanalgorithm.
https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm
}

\keyword{ GCD, Euclidean Algorithm }
\seealso{https://en.wikipedia.org/wiki/Greatest_common_divisor.
https://www.youtube.com/watch?v=gdmfOwyQlcI.}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
dijkstra(wiki_graph, 1)
dijkstra(wiki_graph, 3)
}
