\name{RCourseLab-package}
\alias{RCourseLab-package}
\docType{package}
\title{
\packageTitle{RCourseLab}
}
\description{
\packageDescription{Euclidean() returns the greatest common divisor (GCD) of two numbers.}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Euclidean() returns the greatest common divisor (GCD) of two numbers.}
\packageIndices{RCourseLab}
If input data is numeric or integer then 
the output will be a integer otheriwise it will return an error.
Euclidean() returns the greatest common divisor (GCD) of two numbers.
}
\author{
\packageAuthor{ Authors@R: c(
    person("Farhana", "Chowdhury Tondra", email = "farch587@student.liu.se", role = "cre"),
    person("Farha", "Shazmeen", email = "syesh076@student.liu.se", role = "aut"))}

Maintainer: \packageMaintainer{ tondralok@yahoo.com}
}
\references{
https://en.wikipedia.org/wiki/Euclidean algorithm.
}

\keyword{ GCD, Euclidean Algorithm }
\seealso{

\code{\link[<pkg>:<pkg>-package]{<pkg>}}}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
