% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCourseLab.R
\docType{package}
\name{RCourseLab}
\alias{RCourseLab}
\alias{RCourseLab-package}
\title{A package for Lab 3 Assigments.}
\description{
Package contains of implementations of two algorithm.
Dijkstra is for finding the shortest path from the initial node and Euclidean is for calculating greatest common divisior of two numbers.
This package provides two functions:
euclidean & dijkstra.
}
\section{euclidean functions}{

Euclidean() returns the greatest common divisor (GCD) of two numbers.
}

\section{dijkstra functions}{

dijkstra() returns the shortest path based on the algorithm.
}

\seealso{
\link[RCourseLab]{euclidean}

\link[RCourseLab]{dijkstra}

\link[RCourseLab]{wiki_graph}
}
